requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
requests
beautifulsoup4
selenium
webdriver-manager
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.0